#325 initial refactor work
