+++ functionary/ui/templates/core/workflow_detail.html
@@ -1,4 +1,5 @@
 {% extends "base.html" %}
+{% load extras %}
 {% block content %}
     <nav aria-label="breadcrumb">
         <ol class="breadcrumb my-1 pb-1">
@@ -11,32 +12,64 @@
     <hr />
     <div class="row">
         <div class="col-10 col-md-8 col-xl-7">
-            <div class="hstack gap-3 mt-2 mb-1">
-                <h2>
-                    <i class="fa fa-diagram-next"></i>
-                    {{ workflow.name }}
-                </h2>
-                <i class="fa fa-pencil-alt fa-lg text-info"
-                   type="button"
-                   title="Edit"
-                   hx-target="body"
-                   hx-swap="beforeend"
-                   hx-get="{% url 'ui:workflow-update' workflow.id %}?id={{ workflow.id }}"></i>
-            </div>
+            <h2 class="mt-2 mb-1">
+                <i class="fa fa-diagram-next"></i>
+                <span>{{ workflow.name }}</span>
+            </h2>
+            {% if workflow.summary %}<p class="mb-1 text-muted">{{ workflow.summary }}</p>{% endif %}
             <div class="vstack gap-3 my-4">
-                {% if workflow.description %}<div id="desc" class="ms-4">{{ workflow.description|linebreaks }}</div>{% endif %}
-                <div>
-                    <h3 class="fw-semibold">
-                        <i class="fa fa-list fa-sm fa-fw me-1"></i>Parameters
-                    </h3>
-                    {% include "partials/workflows/parameter_list.html" %}
-                </div>
-                <div>
-                    <h3 class="fw-semibold">
-                        <i class="fa fa-stairs fa-sm fa-fw me-1"></i>Steps
-                    </h3>
-                    {% include "partials/workflows/step_list.html" %}
-                </div>
+                {% if workflow.description %}<div>{{ workflow.description | linebreaks }}</div>{% endif %}
+                {% if form is not None %}
+                    {% if workflow.active %}
+                        {% if missing_variables %}
+                            <div class="alert alert-danger" role="alert">
+                                <p>Tasking this workflow may not work as expected; the following variables are missing:</p>
+                                <ul>
+                                    {% for missing in missing_variables %}<li>{{ missing }}</li>{% endfor %}
+                                </ul>
+                            </div>
+                        {% endif %}
+                        <div class="border border-dark-subtle rounded form-fill">
+                            {% if form.fields %}
+                                <div class="row mx-2 mt-3">
+                                    <div class="col col-lg-10 col-xl-9 col-xxl-8">
+                                        <form id="taskCreateForm"
+                                              method="post"
+                                              action="{% url 'ui:function-execute' %}"
+                                              enctype="multipart/form-data">
+                                            {% csrf_token %}
+                                            <input type="hidden" name="workflow_id" value="{{ workflow.id }}"/>
+                                            {{ form.non_field_errors }}
+                                            {{ form }}
+                                        </form>
+                                    </div>
+                                </div>
+                            {% else %}
+                                <p class="font-monospace m-4">No parameters defined.</p>
+                            {% endif %}
+                        </div>
+                        {% if workflow.first_step %}
+                            <div>
+                                <button class="btn {% if missing_variables %}btn-warning{% else %}btn-primary{% endif %}"
+                                        type="submit"
+                                        title="Execute"
+                                        form="taskCreateForm">
+                                    Execute
+                                </button>
+                            </div>
+                        {% else %}
+                            <div class="alert alert-info text-center" role="alert">
+                                <i class="fa fa-circle-info"></i>
+                                This workflow has no steps, so it cannot be executed.
+                            </div>
+                        {% endif %}
+                    {% else %}
+                        <div class="alert alert-warning text-center" role="alert">
+                            <i class="fa fa-triangle-exclamation"></i>
+                            This workflow is currently inactive.
+                        </div>
+                    {% endif %}
+                {% endif %}
             </div>
         </div>
     </div>
