--- functionary/ui/templates/core/workflow_task.html
@@ -1,76 +1,0 @@
-{% extends "base.html" %}
-{% load extras %}
-{% block content %}
-    <nav aria-label="breadcrumb">
-        <ol class="breadcrumb my-1 pb-1">
-            <li class="breadcrumb-item">
-                <a href="{% url 'ui:workflow-list' %}">Workflows</a>
-            </li>
-            <li class="breadcrumb-item active" aria-current="page">{{ workflow.name }}</li>
-        </ol>
-    </nav>
-    <hr />
-    <div class="row">
-        <div class="col-10 col-md-8 col-xl-7">
-            <h2 class="mt-2 mb-1">
-                <i class="fa fa-diagram-next"></i>
-                <span>{{ workflow.name }}</span>
-            </h2>
-            {% if workflow.summary %}<p class="mb-1 text-muted">{{ workflow.summary }}</p>{% endif %}
-            <div class="vstack gap-3 my-4">
-                {% if workflow.description %}<div>{{ workflow.description | linebreaks }}</div>{% endif %}
-                {% if form is not None %}
-                    {% if workflow.active %}
-                        {% if missing_variables %}
-                            <div class="alert alert-danger" role="alert">
-                                <p>Tasking this workflow may not work as expected; the following variables are missing:</p>
-                                <ul>
-                                    {% for missing in missing_variables %}<li>{{ missing }}</li>{% endfor %}
-                                </ul>
-                            </div>
-                        {% endif %}
-                        <div class="border border-dark-subtle rounded form-fill">
-                            {% if form.fields %}
-                                <div class="row mx-2 mt-3">
-                                    <div class="col col-lg-10 col-xl-9 col-xxl-8">
-                                        <form id="taskCreateForm"
-                                              method="post"
-                                              action="{% url 'ui:function-execute' %}"
-                                              enctype="multipart/form-data">
-                                            {% csrf_token %}
-                                            <input type="hidden" name="workflow_id" value="{{ workflow.id }}"/>
-                                            {{ form.non_field_errors }}
-                                            {{ form }}
-                                        </form>
-                                    </div>
-                                </div>
-                            {% else %}
-                                <p class="font-monospace m-4">No parameters defined.</p>
-                            {% endif %}
-                        </div>
-                        {% if workflow.first_step %}
-                            <div>
-                                <button class="btn {% if missing_variables %}btn-warning{% else %}btn-primary{% endif %}"
-                                        type="submit"
-                                        title="Execute"
-                                        form="taskCreateForm">
-                                    Execute
-                                </button>
-                            </div>
-                        {% else %}
-                            <div class="alert alert-info text-center" role="alert">
-                                <i class="fa fa-circle-info"></i>
-                                This workflow has no steps, so it cannot be executed.
-                            </div>
-                        {% endif %}
-                    {% else %}
-                        <div class="alert alert-warning text-center" role="alert">
-                            <i class="fa fa-triangle-exclamation"></i>
-                            This workflow is currently inactive.
-                        </div>
-                    {% endif %}
-                {% endif %}
-            </div>
-        </div>
-    </div>
-{% endblock content %}
