+++ functionary/ui/templates/forms/workflow/parameter_edit.html
@@ -6,7 +6,7 @@
 {% block modal_content %}
     <form id="parameter-form"
           method="post"
-          hx-post="{% if workflowparameter %}{% url 'ui:workflowparameter-edit' workflow.pk workflowparameter.pk %}{% else %}{% url 'ui:workflowparameter-create' workflow.pk %}{% endif %}"
+          hx-post="{% if workflowparameter %}{% url 'ui:workflowparameter-update' workflow.pk workflowparameter.pk %}{% else %}{% url 'ui:workflowparameter-create' workflow.pk %}{% endif %}"
           hx-target="#form-modal"
           hx-swap="outerHTML"
           enctype="multipart/form-data">
