+++ functionary/ui/views/workflow/step/update.py
@@ -55,7 +55,7 @@ class WorkflowStepUpdateView(PermissionedUpdateView):
             step_form.instance.parameter_template = parameter_form.parameter_template
             step = step_form.save()
 
-            success_url = reverse("ui:workflow-detail", kwargs={"pk": step.workflow.pk})
+            success_url = reverse("ui:workflow-update", kwargs={"pk": step.workflow.pk})
 
             return HttpResponseClientRedirect(success_url)
         else:
