+++ functionary/ui/views/workflow/update.py
@@ -1,4 +1,4 @@
-from django.urls import reverse
+from django.urls import reverse_lazy
 from django_htmx.http import HttpResponseClientRedirect
 
 from core.models import Workflow
@@ -14,6 +14,6 @@ class WorkflowUpdateView(PermissionedUpdateView):
 
     def form_valid(self, form):
         form.save()
-        success_url = reverse("ui:workflow-detail", kwargs={"pk": form.instance.pk})
+        success_url = reverse_lazy("ui:workflow-list")
 
         return HttpResponseClientRedirect(success_url)
